refactored App.js form element into its own component :link::link:
