+++ src/components/Form.css
@@ -1,0 +1,10 @@
+.input-container {
+  display: flex;
+  flex-direction: column;
+  width: 300px;
+}
+
+.form {
+  display: flex;
+  gap: 2em;
+}
